Split MarkdownLanguage types into separate files
