+++ Sources/SwiftParser/Languages/MarkdownLanguage+ParagraphBuilder.swift
@@ -1,0 +1,56 @@
+import Foundation
+
+extension MarkdownLanguage {
+    public class ParagraphBuilder: CodeElementBuilder {
+        public init() {}
+        public func accept(context: CodeContext, token: any CodeToken) -> Bool {
+            if token is Token { return true } else { return false }
+        }
+        public func build(context: inout CodeContext) {
+            var tokens: [Token] = []
+            var ended = false
+            while context.index < context.tokens.count {
+                guard let tok = context.tokens[context.index] as? Token else { context.index += 1; continue }
+                switch tok {
+                case .text, .star, .underscore, .backtick:
+                    tokens.append(tok)
+                    context.index += 1
+                case .hardBreak:
+                    while let last = tokens.last, case .text(let s, _) = last, s.allSatisfy({ $0 == " " }) {
+                        tokens.removeLast()
+                    }
+                    tokens.append(tok)
+                    context.index += 1
+                case .newline:
+                    context.index += 1
+                    ended = true
+                case .dash, .hash, .plus, .lbracket,
+                     .greaterThan, .exclamation, .tilde, .equal, .lessThan, .ampersand, .semicolon, .pipe:
+                    ended = true
+                case .number:
+                    if context.index + 1 < context.tokens.count,
+                       let dot = context.tokens[context.index + 1] as? Token,
+                       case .dot = dot {
+                        ended = true
+                    } else {
+                        tokens.append(tok)
+                        context.index += 1
+                    }
+                case .eof:
+                    context.index += 1
+                    ended = true
+                case .dot, .rbracket, .lparen, .rparen:
+                    tokens.append(tok)
+                    context.index += 1
+                }
+                if ended { break }
+            }
+
+            let value = tokens.map { $0.text }.joined()
+            let children = MarkdownLanguage.parseInlineTokens(tokens, input: context.input)
+            let node = MarkdownParagraphNode(value: value)
+            children.forEach { node.addChild($0) }
+            context.currentNode.addChild(node)
+        }
+    }
+}
